+++ pkg/commands/apisecurity/discoveredoperations/doc.go
@@ -1,0 +1,2 @@
+// Package discoveredoperations contains commands to list and update discovered operations.
+package discoveredoperations