+++ pkg/commands/apisecurity/operations/doc.go
@@ -1,0 +1,2 @@
+// Package operations contains commands to manage operations associated with services.
+package operations